!11 Resolve internal issue #331: output_verification function fails on NaN After fixing internal issue #313 the CG algorithm silently failed due to the presence of NaN. This MR fixes the `output_verification` function in `test/utils` to ensure verification fails in such cases. This is done by use of cmath's comparison functions that are guaranteed to behave as prescribed by the IEEE floating point standard and by reviewing the verification logic, applying fixes where needed.
Most fixes that are not comparison-related end up moving assertions that could fail in the presence of NaNs, to if-statements that, if triggered, generate new error codes. This MR retains the behaviour that all possible errors that can be detected are both detected and reported, as opposed to immediately returning when one error is found.

This MR also adds more verbose output even on verification success, always checks that the vector output by ALP/GraphBLAS is indeed dense (not just when assertions are enabled), and adds documentation to the function.

Some minor coding style fixes are included also.