!15 Resolve internal issue #352: fix bug in argmin and argmax The out-of-place application of both the argmin and argmax operation was bugged. This MR fixes the issue, and strengthens the corresponding unit tests to be able to detect this and similar issues.
Expected impact of this bug is low since the use of argmin outside of tests depends on the in-place application, while on the other hand there is no use of argmax outside of the test suite. This fix will hence be folded into the next regular release.

This MR also includes various minor code style fixes.