!21 Resolve internal issue #355: default and modifiable capacities for all ALP/GraphBLAS containers This MR introduces default capacities for vectors and matrices that are derived from the container size. It furthermore introduces `grb::capacity` to query the current capacity (in number of nonzeroes or values), that a container currently contains. Capacities now also can be requested by passing a minimum required capacity to the container constructor. As before, capacities furthermore can be modified after containers have been instantiated via a call to `grb::resize`.
This MR also introduces memory storage requirements as part of the semantics, and adapts the documentation for `grb::{size,resize,capacity,nnz,clear}` as well as the documentation of the `grb::{vector,matrix}` constructors accordingly. This allows for the new `grb::{resize,capacity}` functionality to be used in memory-constrained applications-- both in cases that arise from target system limitations and/or from particularly demanding scales of input data.

The documentation for aforementioned functions now furthermore is brought up to date with the idea that backends can attach different performance semantics to functions. Exceptions are  _getters_ , which have very strict constant-time performance semantics that all backends must comply with. Such getters are now also explicitly indicated as such in the base specification.

Next, all algorithms are modified to, if they assume a certain capacity, to check that the given outputs and buffers have that capacity-- and otherwise return `ILLEGAL`. The documentations of all functions have been revised.

This MR adds a unit test for the new `grb::capacity` and associated changed semantics, and refactors the allocation mechanism of reference matrices.

Finally, this MR includes a couple of unrelated bugfixes and improvements. The following issues are now fixed:
1. label propagation's convergence detection was broken;
2. label propagation's smoke test did not properly check the number of iterations taken;
3. some primitives that should be `noexcept` were not;
4. the documentation of `graphblas.hpp` was unclear about the order of includes it enforces;
5. some algorithms required users pass (structurally) dense I/O arguments, while it could "densify" sparse arguments when required;
6. some algorithms did not specify their behaviour fully, in particular in terms of performance semantics and which error codes could be returned;
7. the k-hop BFS (or k nearest neighbours) required two buffers, while only one was necessary;
8. no centralised allocations of all shared and private memory regions in reference matrix construction;
9. pure I/O routines on ALP/GraphBLAS containers were scattered across `blas{1,2,3}.hpp` while they would better fit in `io.hpp` instead;
10. in `reference/blas3.hpp` an internal getter was used for write access;
11. some unit tests when fail could be quite verbose, instead of redirecting `stderr` to a log file;
12. improved interface of the sparse/graph neural network inference algorithm;
13. some read-only fields were not annotated as `const`;
14. constructing an empty BSP1D matrix leaved uninitialised process-local matrices. 

No new unit test was added regarding the latter issue, as the new capacity unit test is able to detect regression.

As always, the MR also includes code style fixes across the files it touches.