!22 Resolve internal issue #374: MatrixVectorIterator copies source vector and may destroy the copy before iterating over it This MR refactors the MatrixVectorIterator to not propagate vectors to base Converter classes, but rather propagates vector start and end iterators. It also reduces the number of template arguments that propagate, and updates the associated unit test.
The bug this MR fixed was observed via the use of `grb::outer`. This primitive had an additional bug where out-of-place behaviour was assumed but not checked. A future issue will resolve that the outer product should be calculated in-place and over a semiring.

As always, this MR applies various code style fixes throughout all files touched while resolving this issue.