!34 Resolve internal issue #322: Rename phases, and add phases to all primitives with potentially sparse output Prior to this MR, phases referred to level-3 operations only, and where modeled after the classical two-stage symbolic and numeric approach. To also incorporate one-phase approaches as well as two-stage approaches based on estimation rather than the full simulation and precise counting of fill-in during a "symbolic" step, and following the MR !21 that introduced container capacities, we rename the phases to RESIZE and EXECUTE.
The RESIZE phase does not modify the contents of containers but could increase its capacity as necessary to hold the output of the requested computation. Different from `grb::resize`, shrinking capacities by RESIZE phases is not allowed.

The EXECUTE phase assumes the capacity of output containers is sufficient and attempts to compute the full output. If it turns out the capacity was in fact insufficient, the new semantics introduced with this MR will cause the primitive to return `FAILED`. On failure, the output container will be cleared.

Backends must define performance semantics across the usual dimensions for each primitives in each phase. Note that since MR !21, performance semantics include specifying when changes to the memory usage may occur, and if so, includes specifying by how much.

Finally, this MR introduces an optional Phase argument to *every* ALP/GraphBLAS primitive that has vector or matrix output; i.e., it no longer only applies to level-3 primitives. If the output is not a vector or matrix container, such as e.g. for `grb::dot`, then no Phase argument may be given. The default Phase argument for primitives that have it is `EXECUTE`.

This MR implements the revised semantics throughout the `reference`, `reference_omp`, `bsp1d`, and `hybrid` backends. It furthermore specifies an experimental `TRY` phase, but does not implement it; exploring the use of this phase will be future work.

Additional changes due to related refactoring:
1. Since this MR touches all `blas{1,2,3}.hpp` files in the base and all backend directories, we took this opportunity to also resolve all code style issues that were introduced by clang-format which got horribly confused primarily by the double-compilation trick used in the `reference` backend (#I43XSD ). These files represented the bulk of affected sources affected by issue #I43XSD.
2. The definition of the `norm2` function has moved to `algorithms`