!37 Resolve internal issue #387: introduce grb::wait This MR adds a `grb::wait` primitive that is useful for backends that support the nonblocking mode, such as the one described by Mastoras et al. [1]. In a non-blocking mode, ALP/GraphBLAS primitives may be lazily evaluated. A sequence of such unexecuted primitives must execute at so-called trigger points, such as when the user requests an output iterator, causing data from opaque containers to be extracted from the control of ALP/GraphBLAS. The `grb::wait` allows manually triggering all pending operations.
This MR includes a trivial implementation for the current backends, which are all blocking-- i.e., for all current backends, the the `grb::wait` does nothing except return `grb::SUCCESS`. The MR also introduces a unit test for `grb::wait` that is designed for both blocking and non-blocking backends.

As always, the MR includes some minor code style fixes.

[1] http://albert-jan.yzelman.net/publications.php#mastoras22