!39 Resolve internal issue #384: bsp1d/blas1.hpp dead code removal and bugfixes Prior to this MR, the following code appeared in `bsp1d/blas1.hpp`:
```
        RC foldr(
                const InputType &alpha,
                Vector< IOType, BSP1D, Coords > &y,
                const Monoid &monoid,
                ...
        ) {
                // simply delegate to reference implementation will yield correct result
                RC ret = foldr< descr >( alpha, internal::getLocal( y ), monoid,
                        phase );
                if( ret == SUCCESS ) {
                        internal::getCoordinates( internal::getGlobal( y ) ).assignAll();
                }
                return ret;
        }
```

This is odd because the `assignAll` should be done by the `foldr` it delegates to. Instead, the code should ensure that the global BSP1D vector knows that it has become dense.

A review of the entire file found one other occurrence and found several similar issues. This MR fixes all of them. It also adds static checks to the top-level BSP1D implementation rather than relying on the process-local backend to emit them, since the latter option results in less readable error messages.

The review showed that failed EXECUTE and failed TRY phases were not always handled correctly. This MR fixes those issues for EXECUTE as well, and fixes some of those issues for TRY. (Note that the difficulty for TRY is that no process-local backends have implemented that phase yet.)

Finally and similarly to the above, it is not always correct to let process-local backends determine whether a call should return ILLEGAL. Therefore, this MR modifies the current behaviour to make such dynamic checks prior to delegating to process-local backends.

N.B.: Deprecated methods were not checked thoroughly.