!42 Resolve internal issue #394: insufficient buffer for eWiseLambda on matrices could occur The reference `grb::eWiseLambda` on matrices requires a buffer to store possibly updated nonzeroes as it executes the lambda on the column-major storage, thus requiring transposition into the row-major storage. This proceeds by copying the CCS updates into a (thread-)local buffer of limited size, and then bulk-updating the CRS store once the local buffer is full.
The minimum size of the thread-local buffer is `T*L*sizeof(D)` where `T` is the number of threads, `L` the cache line size (in bytes), and `D` the nonzero type of the matrix nonzeroes. This minimum buffer size was not enforced by the matrix constructor, which this MR fixes.