Resolve internal issue 411: create a transition path sparseblas target (#11)
This MR introduces a proof-of-concept where ALP/GraphBLAS is used to generate a (subset) of the standard blas_sparse.h and spblas.h SparseBLAS APIs. The former is the SparseBLAS standard as proposed by NIST and the BLAS forum. The latter is a de-facto standard that is based on a pre-standard version of the NIST SparseBLAS. With this merge, CMake will install the spblas and blas_sparse headers in the `include/transition` directory, and will install `libsparseblas` and `libsparseblas_omp` into `lib/sequential`; that is, we generate both a sequential as well as a shared-memory parallel version of both SparseBLAS APIs.

This is dubbed a "transition path" since it allows the use of software as-is currently implemented based on a standard API without having to re-write (parts of) the user code to follow the ALP syntax -- even though the latter may be simple to do. For the SparseBLAS API this does not come at any noticeable overhead for the currently implemented functions, while for the SpBLAS API some overhead is incurred for the matrix--vector and matrix--matrix multiplication routines. This is due to the transparent matrix type in spblas, versus an opaque one in the standard SparseBLAS.

This does *not* mean that the performance is totally on par with ALP/GraphBLAS, even when using the standard API; for example, the optimisations performed by the nonblocking backend of ALP/GraphBLAS cannot be applied in the transition path (unless a JIT approach is followed).

The development history of this proof-of-concept may be viewed in the branch associated to Github PR #11 or the branch associated to internal issue 411 on Gitee.

This MR also:
 - the reference matrix output iterator could enter an infinite loop and go out-of-bounds, herewith fixed;
 - the reference grb::eWiseApply ([T1]<-T2<-T3), monoid variant would EXECUTE instead of RESIZE when the latter was asked for, herewith fixed;
 - the reference allocator may, for large allocations, print an ill-formatted string, herewith fixed;
 - adds a SparseBLAS extension for opaque vector types, adds a sparse matrix--sparse vector multiplication using that type, adds a sparse matrix--sparse matrix multiplication for the SparseBLAS standard (not for spblas, as one was already commonly defined there), and adds sparse vector and matrix read-out routines;
 - ensured compilation with LLVM Clang (though this remains not "officially" supported at this stage);
 - fixes various code style issues.