Resolves internal issues 441 and 311: missing run time checks to confirm density when dense descriptor is given to foldl (vector to scalar) (#19)
Aristeidis detected that when giving a dense descriptor and a sparse vector to `grb::foldl` (vector to scalar reduction), that the resulting code will complete the requested computation. It was instead intended, like with other primitives, that `ILLEGAL` is returned instead since the programmer promises a certain property that does not hold in practice.

This is conform the intent that the dense descriptor 1. prevents the generation of code that handles sparse cases, and 2. achieves (minor) performance benefits as it can skip some of the associated checks for sparsity.

This MR ensures the intended behaviour is what occurs. It also:
- reduces some code duplication throughout the code base;
- removes grb::Phase as an argument to reduction to a scalar, since the output is not a container;
- revises and updates the documentation regarding `grb::{foldl,foldr}` (vector to scalar variants);
- introduces an input-masked `grb::foldr` variant. Input masking existed for the `grb::foldl` in order to support HPCG, but introducing a similar input mask for the foldr appeared to have been overlooked, herewith fixed;
- introduces tests for the input-masked foldr;
- introduces tests for when ILLEGAL should be returned;
- introduces tests for when MISMATCH should be returned;
- simplifies the reduction of a vector to scalar unit test by removing the (duplicate) performance test it had embedded;
- increase unit tests coverage with non-trivial masks and `descriptors::invert_mask`.

As always, also included are some minor code style and typo fixes across the files this MR touches.

(The commit history of this MR is retained in branch 441-missing-run-time-checks-to-confirm-density-when-dense-descriptor-is-given.)