Clarify the dense descriptor (#38)
The documentation of the dense descriptor was incorrect and could do with clarifying corner cases.

This MR adds such documentation, and also moves an exposed internal include from the header file to the corresponding `descriptors.cpp`. It also clarifies the `toString' output of the dense descriptor, and applies minor code style fixes.
 
This MR corresponds to, and addresses, internal issue #485. Thanks to Aristeidis for reporting.