422 clear setelement and usage of pinnedvector leads to segementation fault (#45)
Aristeidis and Anders reported an issue where combinations of `clear`, `setElement`, and usage of the `PinnedVector` could segfault for the `reference` and `reference_omp` backends. This MR first and foremost introduces a unit test for the `PinnedVector` that operates on vectors of `double`s, vectors of `std::pair`s, and tests it using both the SEQUENTIAL and PARALLEL I/O modes.

This added unit test found or confirmed the following issues, which this MR also fixes:
- `grb::set` (BSP1D) contains code to handle the `use_index` descriptor, but this code should only pass by the compiler if `size_t`s are convertible into the vector element type;
- `grb::PinnedVector` constructor (reference, reference_omp, and BSP1D) should pin the stack, not the bitmask;
- `grb::clear` (BSP1D) did not undo the `became_dense` flag, potentially resulting in a distributed vector that is both cleared and became dense;
- `grb::PinnedVector` constructor (BSP1D) now throws expections instead of silently ignoring errors during construction.

Also minor performance enhancements, code style fixes, better debug tracing, code documentation improvements, and a fixed typo.

This MR closes internal issue #422.