Resolve internal issue #137: re-enable overlap detection between output vector and output mask (#52)
In sparse matrix--vector multiplication, there was disabled code detecting whether the output and output mask were the same vector. When considering the scattering SpMV multiplication, which consists of a loop of updates to the output vector, in such an overlapping case the intermediate scatters would overwrite the mask, potentially altering it during the multiplication. Therefore surely this overlap must be illegal (as it also is defined in the code documentation as well as the initial ALP/GraphBLAS paper), and the overlap check must be re-enabled.

This MR makes indeed re-enables the check, and revises the documentation of `grb::mxv` and `grb::vxm` in `base/blas2.hpp` to bring it up to date with the current state, and in particular with regards to the overlap detection and the `descriptors::safe_overlap` descriptor.

It additionally revises the code in `base/blas2.hpp` to incorporate `grb::Phase` arguments, and takes care to only document the full versions of `grb::mxv` and `grb::vxm`, and not the short-hand versions with mask arguments omitted. It also adds some missing SFINAE checks in the file, and moves the (unchanged) mxv and vxm performance semantics to the backends that define them.