512 implement matrix view over a vector (#60)
Taking a matrix view over a vector produces a column matrix of size Mx1, where M is the vector length. In other words, the vector is observed as a column matrix. As a result, vectors may be used as matrices in operations that operate on matrices.
The list of changes this commits introduces is the following:
* Add types to support Matrix view over a vector
* Add AMF reshape specialization for creating a matrix view
* Add get_view that implements a matrix view over a vector
* Extend unit test with the new view type