523 dense file iterator (#70)
Implementation of file reader and iterator for use in buildMatrix() and similar.

This implementation reads only  dense matrices saved in Matrix Market Array Format.

Currently only forward access iterator is supported. 
Matrix Market Array Format stores elements in column-wise order and this iterator is returning elements in the same order.

For matrices with symmetry, Symmetric and  Hermitian matrices are supported, 
but iterator reads and returns only lower triangular part of matrix.
It should be taken care, within ALP code, to build the proper matrix, from this lower triangular part, or implement the corresponding logic.

Buffer for matrix elements has been removed, and should be implemented in the future.

It is suggested that this utility should be merged with analogous reader for sparse matrices, implemented in GRB namespace. 


Co-authored-by: Denis Jelovina <denis.jelovina@huawei.com>