419 benchmarking spmv spmspv spmspm (#75)
This MR adds three kernel-level performance tests to ALP/GraphBLAS: sparse matrix--dense vector multiplication, sparse matrix--sparse vector multiplication, and sparse matrix--sparse matrix multiplication. These three tests run on arbitrary input data sets, and employ the usual benchmarking methodology by default: run at least 10 experiments while computing the min, max, avg, and stddev aggregates and while sleeping for one second between experiments.
Individual timings, again when used in default mode, will take at least 100 ms.-- if a single computation is shorter than that, then an "individual" timing in fact corresponds to n such timings, where n is the minimum number of times for which the computation takes 100 milliseconds or more. Both the resulting "inner" (100 ms.) and "outer" (10) repetitions may of course also be manually set.

The performance test script is updated to run these three new performance tests on a fixed series of datasets. If one of the test sets is not available, then the experiment is skipped. For the sparse matrix--sparse matrix multiplication, square matrices are multiplied with itself; this test is, by default, only enabled for the smaller matrices in the dataset. The dataset series can easily be extended in the script should need be.

The new performance tests are only run for the smallest datasets for the hyperdags backend, in-line with the current behaviour of performance tests for that backend already present in the main branch.

The MR also includes minor code style fixes, as well as a workaround for CIs that run out of storage capacity due to the -g generating large binary files, while, related, not archiving some additional files that are not required to execute tests after building them.

Thanks to Anders Hansson and Aristeidis Mastoras for contributing this MR!