536 copying vector by assignment multiple times fails asserts (#85)
Bugfix for the following sequence that previously triggered an assertion failure for the `reference_omp` backend:
```
Vector<double> a(n), b(n);
set(a, 1.2);
set(b, 1.5);
a=b;
a=b; //Fails assertion
```

Issue: multiple-registration of the same container.

Cause: the previous implementation of the copy-assignment seems to predate the introduction of IDs. In its implementation, the container `a` is in fact replaced by a totally new container, with a different ID, through the use of `std::swap` and the creation of a new container within the copy-assignment. Additionally, this induces new memory allocations while in principle a dispatch to `grb::set` would have sufficed.

Solution: dispatch to `grb::set` instead. Apart from the performance gain, this also retains the container's original ID and does not trigger the earlier double-registration issue and thus does not trigger the reported assertion failure. Additionally, this MR adds a unit test for double copy-assignment.

This MR also includes
- bugfix: BSP1D `grb::set` was not correctly updated as part of the previous MR (internal issue 533);
- implement previously missing copy-assignment for BSP1D backend;
- improved `_DEBUG` tracing;
- minor code style fixes.

Reported by Anders Hansson -- many thanks!