Resolve internal issue 544: bug in ewisemuladd dispatch (#110)
The mask-driven variant of `eWiseMulAdd` was never dispatched to when using scalar inputs, while its implementation was bugged in those cases. While the primitive is deprecated, this bug is still relevant because most variants of `eWiseMul` in fact dispatch to `eWiseMulAdd`.

Perhaps taking the difficult route, this MR (finally) reviews the implementation of the deprecated `eWiseMulAdd` and `eWiseAdd`. The former is fully reviewed (since it is used for the non-deprecated eWiseMul), while the latter only has its functional behaviour reviewed. The `eWiseMulAdd` variants as well as its dispatcher have been optimised to deal with the specific case of `eWiseMul`.

A new unit test is introduced for `eWiseMul`, testing all its variants. The pre-existing tests for `eWiseMulAdd` are retained. No tests will be added for `eWiseAdd` since it is deprecated.

This MR additionally addresses the following:
- some calls to `eWiseMulAdd` with sparse vectors could trigger invalid assertion failures when compiled without `NDEBUG`;
- using sparse vectors with the dense descriptor in `eWiseMul` and `eWiseMulAdd` did not return `ILLEGAL`, now fixed;
- not all variants of `eWiseMul` were provided in all backends, now fixed;
- the unmasked `eWiseMul` with all vector inputs was erroneously registered as `eWiseMulAdd` in the hyperdags backend, now fixed;
- some variants of `eWiseAdd` and `eWiseMulAdd` remained out-of-place, now fixed;
- improved `_DEBUG` tracing across the functions touched;
- code style fixes across both related and unrelated files.

Thanks to Aristeidis for pointing out the initial issue for the mask-driven eWiseMulAdd with one or more scalar inputs!