Resolves internal issue #577: two bugs related to eWiseApply (#115)
Aristeidis detected that the unit test for `eWiseApply` could fail for some vector lengths that `unittests.sh` never calls it with, and also detected that the sparse-vector driven variant of `eWiseApply` did not correctly check for output mask conditions. The underlying issues were a metabug in the unit test that did not deal properly with vector lengths whose half is an odd number, and a missing negation of the evaluated mask. This MR fixes both issues, and also:
- executes the `eWiseApply` unit test for vectors of length `14`, thus triggering the detecting meta-bug should it ever regress;
- adds to sub-tests to the `eWiseApply` unit test designed to trigger the sparse-vector driven variant with a non-trivial mask that would detect erroneous mask evaluations, thus triggering detection of the main bug this MR fixes;
- improved the output of the `eWiseApply` unit tests; and
- fixes an unrelated code style issue (incorrect indentation).