134 additional internal ci failures triggered for hyperdag in debug mode (#135)
Internal CI detects errors in the hyperDAGs backend for the `eWiseApply` and `zip` unit tests, as well as in the `hpcg` smoke test, only when run in debug mode (without `-DNDEBUG`).

Issue: the debug mode detects disconnected hyperDAGs when they are dumped. This indicates a vertex from which there is no path to some output.

Cause: some variants of `eWiseApply` did not connect to the output container as an input to the `eWiseApply`. This sounds paradoxical since this primitive is out of place, but like for the `clear`, the container must still be registered as a source because its previous contents are invalidated. The same issue existed for `unzip`.

This MR fixes both issues, and fixes the code style of the `tests/unit/zip.cpp` unit test.

Also resolves internal issue 597.