565 k nn and mpv algorithms gives wrong results (#104)
Anders detected an off-by-one error in the mpv (matrix powers) kernel / algorithm, which this MR fixes. This affects the k-NN algorithms as well as the smoke and performance tests based on it, which were previously verified against the sequential ALP/GraphBLAS baseline. This MR also hardens these tests by comparing against a ground truth verified by an external framework, in this case Octave.
This MR has been thoroughly tested using the full test suite. Additionally, an issue (GitHub #136) has been raised to consider an mpv implementation which is not unrolled and simply based on `std::swap`-- suspected is that the current implementation predates supporting swap, and, different from an implementation using `grb::set` instead of swap, a major performance loss compared to an unrolled algorithm no longer is expected.

Both the bug detection and the fix contribution were by Anders -- many thanks!

Co-authored-by: Anders Hansson <anders.hansson@huawei.com>
Co-authored-by: Albert-Jan N. Yzelman <albertjan.yzelman@huawei.com>