555 k core decomposition algorithm (#153)
This MR adds the k-core decomposition algorithm as published by Li et al. (HPEC '21). The implementation has been provided and tested by Anders Hansson -- many thanks for the contribution!

The newly added algorithm has two variants -- one with an OMP critical section within an `eWiseLambda` (as in the original MR), and one "pure" GraphBLAS variant. The former variant applies to single-node backends and to OpenMP-based shared-memory parallel backends-- and not to other backends. The latter variant is standard ALP/GraphBLAS and thus works with any backend, including distributed-memory ones. In limited testing the latter variant is also faster, and therefore chosen as the default variant.

The MR includes smoke tests of both variants on one specific dataset, with the result compared to a ground-truth that was manually compared against an external framework. Thus the k-core algorithm has automatised basic verification for this and future ALP/GraphBLAS releases, and is guarded against functional regressions.

This MR also adds the following:
 - re-introduces the less-than, less-or-equal (leq), greater-than, greater-or-equal (geq) operators;
 - the BSP1D masked foldl from a scalar into a vector were previously not provided, neither for operators nor (unrelated to this MR) for monoids;
 - (unrelated) the BSP1D foldr from a scalar into a vector using an operator was previously not provided.

This MR also fixes the following bugs:
 - one variant of the masked `eWiseApply` for the `reference_omp` backend (mask-driven, Theta(n) loop) may previously not compute the correct output;
 - (unrelated) the BSP1D foldl from a scalar into a vector using operators erroneously did not accept a sparse destination vector.

As always, the MR includes code style fixes across all files touched, as well as minor code style fixes in some unrelated files.

---------

Co-authored-by: Anders Hansson <anders.hansson@huawei.com>
Co-authored-by: Albert-Jan N. Yzelman <albertjan.yzelman@huawei.com>