642 separating built from tested backends (#200)
This MR solves the issue encountered by the internal CI that runs separate test jobs for testing separate backends, including those that require LPF. The build system, by this MR, separates the list of backends that are to be tested, from the list of backends that should be built. Previously there was no distinction, and so one could request the testing of a backend whose dependences were not built. This is what happened in the internal CI just prior to this MR.

After this MR, and as an example, if the user selects the hybrid backend for testing and excludes all other backends for testing, then
1. the reference_omp backends is built because of the hybrid backend depending on the reference_omp backend, and
2. the reference backend is built because of the reference_omp backend depending on the reference backed. However,
3. only the tests for the hybrid backend are built and run.

In some sense, this MR makes the test build system aware of how backends are composed on top of one another.

Thanks a lot to @byjtew (Benjamin Lozes) for providing the code for this fix, and thanks also to Alberto Scolari (@alberto-scolari) for the untiring work on the testing infra.